requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymysql==1.1.0
presto-python-client==0.8.3
pyhive==0.7.0
thrift==0.16.0
pandas==2.1.3
python-jose==3.3.0
gunicorn==21.2.0
tabulate==0.9.0
pytest==7.4.3
requests==2.31.0
mysql-connector-python==8.0.33 